refactor: move test helper imports to .test-utils.ts files
Test-only helper objects (iptablesSharedTestHelpers,
containerLifecycleTestHelpers, copilotApiResolverTestHelpers,
hostEnvTestHelpers, logDiscoveryTestHelpers) were being imported
directly from production modules by test files. This clutters
the production API surface with test-only symbols.

Create co-located .test-utils.ts re-export files and update all
test imports to use them. The production exports remain (required
for module-internal state access) but tests no longer import
directly from production paths.

Fixes #3968
Fixes #3969
