+++ src/container-lifecycle.test-utils.ts
@@ -1,0 +1,5 @@
+/**
+ * Test-only re-export of internal state helpers from container-lifecycle.
+ * Tests should import from this file, not directly from the production module.
+ */
+export { containerLifecycleTestHelpers } from './container-lifecycle';
