+++ src/copilot-api-resolver.test-utils.ts
@@ -1,0 +1,5 @@
+/**
+ * Test-only re-export of internal helpers from copilot-api-resolver.
+ * Tests should import from this file, not directly from the production module.
+ */
+export { copilotApiResolverTestHelpers } from './copilot-api-resolver';
