+++ src/host-env.test-utils.ts
@@ -1,0 +1,5 @@
+/**
+ * Test-only re-export of internal helpers from host-env.
+ * Tests should import from this file, not directly from the production module.
+ */
+export { hostEnvTestHelpers } from './host-env';
