+++ src/logs/log-discovery.test-utils.ts
@@ -1,0 +1,5 @@
+/**
+ * Test-only re-export of internal helpers from log-discovery.
+ * Tests should import from this file, not directly from the production module.
+ */
+export { logDiscoveryTestHelpers } from './log-discovery';
